changed accelerometer values to ints so I can read from XYZCounts, fixed readings

+++ hardware.h
@@ -7,7 +7,7 @@
  */
 struct GameInputs {
     int b1, b2, b3;     // Button presses
-    double ax, ay, az;  // Accelerometer readings
+    int ax, ay, az;  // Accelerometer readings
 };
 
 /**
